feat: seperate msw handler
#19
