--- src/mocks/handlers/mockUser.ts
@@ -1,4 +1,0 @@
-import {createUser} from '@/mocks/fakers';
-import {Member} from '@/types';
-
-export const mockMembers: Member[] = Array.from({length: 5}, createUser);
